server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
google-genai==1.0.0
httpx==0.28.1
